requirements.txt
beautifulsoup4==4.9.3
dhooks==1.1.4
lxml==4.6.3
Pillow==8.2.0
pytz==2021.1
requests==2.25.1